aws configure 4

+++ workflows/aws_bench_runner.sh
@@ -3,7 +3,9 @@
 # Function to handle errors
 handle_error() {
     echo "An error occurred. Exiting." >&2
-    aws ec2 terminate-instances --instance-ids $instance_id
+    if [ -n "$instance_id" ]; then
+        aws ec2 terminate-instances --instance-ids $instance_id
+    fi
     exit 1
 }
 
@@ -22,7 +24,7 @@ export GITHUB_RUN_NUMBER="$GITHUB_RUN_NUMBER"
 export GITHUB_REF="$GITHUB_REF"
 export READ_GITHUB_ACCESS_TOKEN="$READ_GITHUB_ACCESS_TOKEN"
 
-echo $GITHUB_REF 
+echo $AWS_ACCESS_KEY_ID
 
 # Create an EC2 instance
 instance_id=$(aws ec2 run-instances \
